refactor: DDBEntity lives in gamedata
