+++ src/app/schemas/GameData.ts
@@ -1,0 +1,10 @@
+export class DDBEntity {
+  entity_type: string;
+  name: string;
+  source: string;
+  entity_id: number;
+  page: number | null;
+  is_free: boolean;
+  url: string | null;
+  marketplace_url: string;
+}
