+++ src/app/schemas/Workshop.ts
@@ -79,17 +79,6 @@ export class WorkshopEntitlement {
   required: boolean;
 }
 
-export class DDBEntity {
-  entity_type: string;
-  name: string;
-  source: string;
-  entity_id: number;
-  page: number | null;
-  is_free: boolean;
-  url: string | null;
-  marketplace_url: string;
-}
-
 // helper enums, not exposed in api
 export enum CollectableType {
   ALIAS = 'ALIAS',
